Fix a non-determinism problem
Goals of the form (A, B) where both A and B have alternatives were
not working properly when back-tracking into A.  The alternatives
for B were being lost.  Fix this by creating explicit child
goals for A and B rather than replacing the (A, B) goal with A.
Similarly for (A || B) goals.